python/requirements.txt
Flask==2.1.1
requests==2.26.0
Werkzeug==2.0.3
python-dotenv